Add new_with_shape methods (#700)
Summary:



This diff is to address [this comment](https://www.internalfb.com/diff/D78355743?dst_version_fbid=23914538798238882&transaction_fbid=1255127772923671) from D78355743. Basically the python side wants to use its own logic to slice shape, and build the mesh with the sliced shape it provides.

In the long run, we probably do want to have both Rust and Python layers have their own slicing logic. But the consolidation is not in the scope of this stack.

Differential Revision: D79287947