allow empty workspace + local packaging (#705)
Summary:

for monarch, it is very often that a workspace is not needed, yet a
conda env needs to be specialized. The main function of monarch can be from the
client script. We provide options for users to reuse the workspace from the
image or not providing a workspace at all.

Differential Revision: D79323841