hyperactor_mesh: wire up mesh-managed actors to the global router (#735)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/735

Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/732

When a ProcMesh is created, it is wired into a process-global router, which is used as a fallback for the mesh itself. This is how multiple meshes discover and talk to each other.

However if an actor *in* a mesh creates its own (child) ProcMesh, it is not mututally routable with that child mesh.

This change wires up the various bindings to ensure that they are.

Note that currently this only makes an *actor* that owns a procmesh mutually reachable with the procmesh it creates; it does not yet allow (direct) routing between proc meshes managed by different actors.

Reviewed By: shayne-fletcher, zdevito

Differential Revision: D79478197