ndslice: consistent definition of "len()" (#759)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/759

"len()" should refer to the size of the data structure. In the case of an extent, the natural definition is the number of dimensions. For example, the natural Index implementation for Extent is to give the size of the indexed dimension. This would then require that len() returns the number of dimensions for consistency.

We change 'len' to be defined thus, and provide "num_ranks" to answer how many ranks are contained in the extent.

Reviewed By: shayne-fletcher

Differential Revision: D79569471