"inverted control" mode through Proc::instance
Summary:
Adds Proc::instance() which returns an actor instance and its corresponding handler. This allows the user to create a regular actor without any message handlers. The returned `Instance` provides all the normal capabilities, including sending and receiving messages, being able to spawn and manage child actors, etc.

This is the foundation for a kind of "script mode" actor.

Differential Revision: D79685752