simplify supervision propagation; unhandled events always cause failure (#773)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/773

Currently (local) supervision can propagate events without also killing an intermediate actor.

This is 1) wrong; and 2) complicated.

Instead, we treat an unhandled supervision event as an actor failure, and then reduce the propagation paths to one: that of an actor failing.

In order to retain accurate attribution, we add a "caused_by" field to the actor supervision events.

Reviewed By: shayne-fletcher

Differential Revision: D79702385