Fix supervision error message and printout for `supervision_event` (#803)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/803

Updates the error message for exit from `supervision_event()`.

Why?

The message is clearer and now we can capture a nice stack trace b/c we aren't using the derived Debug of a `SupervisionEvent`, but rather grabbing the actor_id and using the Display of actor_status, which is where the error will be coming from.

Reviewed By: vidhyav, zdevito

Differential Revision: D79810924