Pulling changes of grpo_actor from docs version to original copy (which has edit history).
Summary:
This diff pulls in changes from the doc-team's copy of grpo_actor.py to the original, and removes the copy.  This is to prepare for the next commit, which will move the original (w/ changes) to the doc examples area so we have the same thing but with the change history recovered.

Rollback Plan:

Differential Revision: D80534401