Configurable debounce (#854)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/854

When we increase the number of actors in our simulation it takes longer for all the events at a certain time to complete so we need to wait for longer. If we wait to long then the simulation just runs slower than it needs to so its nice to make this configurable.

In the long term we will come up with a more robust solution to this but in the meantime that is not a priority. See EX528476 to understand the underlying problem the debounce is remedying

Differential Revision: D80137965

Reviewed By: pablorfb-meta