__exit__ in Actors for user defined cleanup
Summary:
There is an open papercut https://github.com/meta-pytorch/monarch/issues/839 where NCCL watchdog causes a segfault on ProcMesh.stop().

This issue can reliably be stopped by making sure `torch.distributed.destroy_process_groups()` is called on exit.

We want to provide users with a convenient way of being able to define a cleanup function for an actor that is automatically invoked on exception or shutdown, and this is a location in which users can add in the call to `destroy_process_groups()`

Because this may not be safe to call (on abnormal exit), we want to be able to also thread through exception details to define different cleanup procedures for normal vs abnormal exit.

When an Actor throws an `Exception` while handling a message, `__exit__` is invoked with the `Exception`. Otherwise, when the Actor exits gracefully it will invoke `__exit__` passing in None as all its arguments

Differential Revision: D80719752