[hyperactor] disallow anonymous Serialized values
We now enforce propagation of type hashes with Serialized. This makes serialized exactly like a serialized version of std::any (i.e., a type-erased, serialized value).

This then allows us to support safe downcasting.

We have this piece of code already, in `mailbox.rs`:

```
        while let Ok(Undeliverable(mut envelope)) = rx.recv().await {
            if let Ok(Undeliverable(e)) = envelope.deserialized::<Undeliverable<MessageEnvelope>>()
            {
                // A non-returnable undeliverable.
                UndeliverableMailboxSender.post(e, monitored_return_handle());
                continue;
            }
```

Previously, this would unconditionally attempt to deserialize the value into an `Undeliverable<MessageEnvelope>`. However this is dangerous, for two reasons: 1) it's possible that we are presented with a differently-typed value that *can* successfully deserialize into this type; 2) the deserializer might go astray. We witnessed the latter, as, with the recent envelope changes, we saw some values cause the deserializer to attempt to deserialize very large values, causing allocator failures.

We should additionally configure our deserializers (bincode, and serde_multipart/bincode) to limit the maximum allowable allocation they are allowed to perform while deserializing.

Differential Revision: [D80955349](https://our.internmc.facebook.com/intern/diff/D80955349/)

ghstack-source-id: 305461156
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/986