Edition 2024 cleanup: rename `expr_2021` -> `expr` (#949)
Summary:

This renames `expr_2021` to `expr` for rust macros. The `expr_2021` was used during the Rust Edition upgrade to ensure compatibility between the two editions. Now that we are using Rust Edition 2024 as the default, and none of the macros are triggering a build failure with this rename, we can rename this back to `expr`

Reviewed By: diliop

Differential Revision: D80508019